% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{add_file_info_attributes}
\alias{add_file_info_attributes}
\title{Helper function: add fileinfo attributes}
\usage{
add_file_info_attributes(obj, path)
}
\arguments{
\item{obj}{object to write attributes to}

\item{path}{path to file}
}
\value{
object with file info attributes
}
\description{
Helper function: add fileinfo attributes
}
