% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_differences.R
\name{check_for_differences}
\alias{check_for_differences}
\title{Check two Dataframes for Differences}
\usage{
check_for_differences(df_x, df_y, dbg = TRUE)
}
\arguments{
\item{df_x}{data frame as retrieved by create_references_df() or
clean_references_df()}

\item{df_y}{data frame as retrieved by create_references_df() or
clean_references_df()}

\item{dbg}{should dbg messages be printed (default: TRUE)}

\item{name_value_x}{column name for values of 'df_x' dataframe (default:
"value_x")}

\item{name_value_y}{column name for values of 'df_y' dataframe (default:
"value_y")}
}
\value{
a dataframe containing only the differences between df_x and df_y
}
\description{
Check two Dataframes for Differences
}
\examples{
\dontrun{
############################################################################
### Option 1
### Check differences between two different versions of "KWB_documents.xml"
############################################################################

old_xml <- extdata_file("2019-01-07_KWB_documents.xml")
new_xml <- extdata_file("2019-01-14_KWB_documents.xml")
old_list <- kwb.endnote::create_endnote_list(old_xml)
new_list <- kwb.endnote::create_endnote_list(new_xml)
old_df <- kwb.endnote::create_references_df(old_list)
new_df <- kwb.endnote::create_references_df(new_list)
diffs_df_oldnew <- check_for_differences(old_df, new_df)
head(diffs_df_oldnew)

############################################################################
### Option 2:
### Check differences between "as-is" import and "collapsing" fields
############################################################################

endnote_list <- create_endnote_list()
refs_df <- create_references_df(endnote_list)
refs_df_collapse <- create_references_df(endnote_list, collapse = TRUE)
diffs_df <- check_for_differences(refs_df, refs_df_collapse)
head(diffs_df)

}
}
