% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_problematic_entries.R
\name{check_problematic_entries}
\alias{check_problematic_entries}
\title{Check Endnote for Problematic References}
\usage{
check_problematic_entries(endnote_list, give_hints = TRUE, dbg = TRUE)
}
\arguments{
\item{endnote_list}{list created with create_endnote_list()}

\item{give_hints}{if TRUE hints will be generated, e.g.
"add_public_or_confidential" for accessiblity data}

\item{dbg}{show debug messages (default: TRUE)}
}
\value{
a data frame with problematic entries
}
\description{
Check Endnote for Problematic References
}
\examples{
\dontrun{endnote_list <- create_endnote_list()
problematic_entries <- check_problematic_entries(endnote_list)
head(problematic_entries)
}
}
