% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_references_df.R
\name{clean_accessibility}
\alias{clean_accessibility}
\title{Helper function: clean access information}
\usage{
clean_accessibility(access, give_hints = FALSE, dbg = TRUE)
}
\arguments{
\item{access}{vector with accessibility information}

\item{give_hints}{if TRUE hints will be generated, e.g.
"add_public_or_confidential" in case of missing entries (default: FALSE)}

\item{dbg}{show debug messages (default: TRUE)}
}
\value{
vector with cleaned accessibility information
}
\description{
Helper function: clean access information
}
