% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_references_df.R
\name{clean_author_names}
\alias{clean_author_names}
\title{Helper function: clean author names}
\usage{
clean_author_names(author_names, give_hints = FALSE, dbg = TRUE)
}
\arguments{
\item{author_names}{with author names to clean}

\item{give_hints}{if TRUE hints will be generated, e.g.
"fix_multiple_authors_per_line" in case of missing entries (default: FALSE)}

\item{dbg}{show debug messages (default: TRUE)}
}
\value{
vector with cleaned author names
}
\description{
Helper function: clean author names
}
