% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_references_df.R
\name{clean_project_names}
\alias{clean_project_names}
\title{Helper function: clean project names}
\usage{
clean_project_names(project_names, give_hints = FALSE, dbg = TRUE)
}
\arguments{
\item{project_names}{with project names to clean}

\item{give_hints}{if TRUE hints will be generated, e.g. "add_project_name"
in case of missing entries (default: FALSE)}

\item{dbg}{show debug messages (default: TRUE)}
}
\value{
vector with cleaned project names
}
\description{
Helper function: clean project names
}
