% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_references_df.R
\name{clean_references_df}
\alias{clean_references_df}
\title{Clean References Dataframe}
\usage{
clean_references_df(references_df)
}
\arguments{
\item{references_df}{as retrieved by create_references_df()}
}
\value{
cleaned references_df
}
\description{
Clean References Dataframe
}
\examples{
\dontrun{endnote_list <- create_endnote_list()
refs_df <- create_references_df(endnote_list)
refs_clean_df <- clean_references_df(refs_df)
head(refs_clean_df)
}
}
