% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KWB_documents.R
\name{create_df_from_endnote_xml}
\alias{create_df_from_endnote_xml}
\title{Create dataframe from Endnote XML file}
\usage{

  create_df_from_endnote_xml(endnote_xml = system.file("extdata/KWB_documents.xml",
  package = "kwb.endnote"))
}
\arguments{
\item{endnote_xml}{path to Endnote library exported as .xml (default:
system.file("extdata/KWB_documents.xml", package = "kwb.endnote")}
}
\value{
data.frame with all information from Endnote XML joined with data from
get_reference_type_names()
}
\description{
Create dataframe from Endnote XML file
}
\examples{
references_df <- create_df_from_endnote_xml()
head(references_df)
}
