% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_endnote_list.R
\name{create_endnote_list}
\alias{create_endnote_list}
\title{Create List From Endnote XML}
\usage{

  create_endnote_list(endnote_xml = system.file("extdata/KWB_documents.xml",
  package = "kwb.endnote"))
}
\arguments{
\item{endnote_xml}{path to Endnote library exported as .xml (default:
system.file("extdata/KWB_documents.xml", package = "kwb.endnote")}
}
\value{
list of imported endnote_xml
}
\description{
Create List From Endnote XML
}
\examples{
references_list <- create_endnote_list()
str(references_list$xml$records[1]$record)
}
