% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_keywords_df.R
\name{create_keywords_df}
\alias{create_keywords_df}
\title{Create Keywords Dataframe}
\usage{
create_keywords_df(references_df)
}
\arguments{
\item{references_df}{references_df as recrieved from
kwb.endnote::create_df_from_endnote_xml()}
}
\value{
keywords dataframe
}
\description{
Create Keywords Dataframe
}
\examples{
references_df <- create_df_from_endnote_xml()
keywords_df <- create_keywords_df(references_df)
head(keywords_df)

}
