% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_list_by_pubtype_from_df.R
\name{create_list_by_pubtype_from_df}
\alias{create_list_by_pubtype_from_df}
\title{Create List By Pub Type From Datafram}
\usage{
create_list_by_pubtype_from_df(refs_df)
}
\arguments{
\item{refs_df}{data frame as created with create_references_df()}
}
\value{
list with references with one sublist for each publication type
}
\description{
Create List By Pub Type From Datafram
}
\examples{
\dontrun{
endnote_list <- create_endnote_list()
refs_df <- create_references_df(endnote_list)
refs_list_by_pubtype <- create_list_by_pubtype_from_df(refs_df)
str(refs_list_by_pubtype, 1)
}
}
