% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_list_with_unique_entries.R
\name{create_list_with_unique_entries}
\alias{create_list_with_unique_entries}
\title{Create List with Unique Entries}
\usage{
create_list_with_unique_entries(refs_df)
}
\arguments{
\item{refs_df}{data frame as created with create_references_df()}
}
\value{
list with unique values for selected columns
}
\description{
Create List with Unique Entries
}
\examples{
\dontrun{endnote_list <- create_endnote_list()
refs_df <- create_references_df(endnote_list)
unique_entries_list <- create_list_with_unique_entries(refs_df)
str(unique_entries_list, 1)
}
}
