% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_references_df_from_list.R
\name{create_references_df}
\alias{create_references_df}
\title{Create References Dataframe}
\usage{
create_references_df(endnote_list)
}
\arguments{
\item{endnote_list}{list created with create_endnote_list()}
}
\value{
data.frame with columns record_id, rec_number, ref_type_id, ref_type_name
}
\description{
Create References Dataframe
}
\examples{
endnote_list <- create_endnote_list()
refs_df <- create_references_df(endnote_list)
head(refs_df)

}
