% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{default_clean_xlsx}
\alias{default_clean_xlsx}
\title{Helper function: default filename for cleaned XLSX}
\usage{
default_clean_xlsx(endnote_list)
}
\arguments{
\item{endnote_list}{list as retrieved by create_endnote_list()}
}
\value{
default clean xlsx filename
}
\description{
Helper function: default filename for cleaned XLSX
}
