% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{default_xlsx}
\alias{default_xlsx}
\title{Helper function: default filename for XLSX}
\usage{
default_xlsx(endnote_list)
}
\arguments{
\item{endnote_list}{list as retrieved by create_endnote_list()}
}
\value{
default xlsx filename
}
\description{
Helper function: default filename for XLSX
}
