% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{default_xml}
\alias{default_xml}
\title{Path to Default XML File}
\usage{
default_xml()
}
\value{
path to xml file stored in this package, containing references from
  KWB Endnote database
}
\description{
Path to Default XML File
}
