% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_references_df_from_list.R
\name{get_authors}
\alias{get_authors}
\title{Helper function: get authors from list for a reference}
\usage{
get_authors(record_list, col_name = "contributors_author")
}
\arguments{
\item{record_list}{list with one record of create_endnote_list()}

\item{col_name}{default: "contributors_author"}
}
\value{
one row authors data frame
}
\description{
Helper function: get authors from list for a reference
}
