% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_available_multi_cols}
\alias{get_available_multi_cols}
\title{Helper Function: Get Available Multi Col Names}
\usage{
get_available_multi_cols(df)
}
\arguments{
\item{df}{as retrieved by create_references_df() or clean_references_df()}
}
\value{
all names which are valid inputs for tidy_selected_cols()
}
\description{
all names which are valid inputs for tidy_selected_cols()
}
