% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_references_df_from_list.R
\name{get_keywords}
\alias{get_keywords}
\title{Helper function: get keywords from list for a reference}
\usage{
get_keywords(record_list, col_name = "keyword",
  extract_value = "keywords")
}
\arguments{
\item{record_list}{list with one record of create_endnote_list()}

\item{col_name}{default: "keyword"}

\item{extract_value}{extract_value = "keyword"}
}
\value{
one row keywords data frame
}
\description{
Helper function: get keywords from list for a reference
}
