% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_references_df_from_list.R
\name{get_pdfurls}
\alias{get_pdfurls}
\title{Helper function: get pdfurls from list for a reference}
\usage{
get_pdfurls(record_list, col_name = "urls_pdf")
}
\arguments{
\item{record_list}{list with one record of create_endnote_list()}

\item{col_name}{default: "url_pdfurls"}
}
\value{
one row pdfurls data frame
}
\description{
Helper function: get pdfurls from list for a reference
}
