% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_reference_type_names}
\alias{get_reference_type_names}
\title{Helper function: get reference type names}
\usage{
get_reference_type_names(endnote_xml = default_xml())
}
\arguments{
\item{endnote_xml}{path to Endnote library exported as .xml (default:
\code{\link{default_xml()}})}
}
\value{
data.frame with columns record_id, rec_number, ref_type_id, ref_type_name
}
\description{
Helper function: get reference type names
}
\examples{
ref_type_names <- get_reference_type_names()
head(ref_type_names)
}
