% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_references_df_from_list.R
\name{get_tertiary_authors}
\alias{get_tertiary_authors}
\title{Helper function: get tertiary authors from list for a reference}
\usage{
get_tertiary_authors(record_list, col_name = "author_tertiary",
  extract_value = "tertiary-authors")
}
\arguments{
\item{record_list}{list with one record of create_endnote_list()}

\item{col_name}{default: "author_tertiary"}

\item{extract_value}{extract_value = "authors"}
}
\value{
one row authors data frame
}
\description{
Helper function: get tertiary authors from list for a reference
}
