% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_xml_filename_without_extension}
\alias{get_xml_filename_without_extension}
\title{Helper function: get xml filename without extension}
\usage{
get_xml_filename_without_extension(obj)
}
\arguments{
\item{obj}{list or datafram as retrieved by create_endnote_list() or
create_df_from_endnote_xml()}
}
\value{
xml filename without file extension
}
\description{
Helper function: get xml filename without extension
}
