% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_references_df.R
\name{give_hints_accessiblity}
\alias{give_hints_accessiblity}
\title{Helper Function: Give Hints For Accessibility}
\usage{
give_hints_accessiblity(access, dbg = TRUE)
}
\arguments{
\item{access}{vector with accessibility metadata to check}

\item{dbg}{should debug messages be printed (default: TRUE)}
}
\value{
vector with access info with hints how to improve data quality (in
case give_hints = TRUE)
}
\description{
Helper Function: Give Hints For Accessibility
}
