% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_hints.R
\name{give_hints_author_names}
\alias{give_hints_author_names}
\title{Helper Function: Give Hints For Author Names}
\usage{
give_hints_author_names(author_names, dbg = TRUE)
}
\arguments{
\item{author_names}{vector with author names to check}

\item{dbg}{should debug messages be printed (default: TRUE)}
}
\value{
vector with author_names with hints how to improve data quality (in
case give_hints = TRUE)
}
\description{
Helper Function: Give Hints For Author Names
}
