% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_references_df.R
\name{give_hints_project_names}
\alias{give_hints_project_names}
\title{Helper Function: Give Hints For Project Names}
\usage{
give_hints_project_names(project_names, dbg = TRUE)
}
\arguments{
\item{project_names}{vector with project names to check}

\item{dbg}{should debug messages be printed (default: TRUE)}
}
\value{
vector with project_names with hints how to improve data quality (in
case give_hints = TRUE)
}
\description{
Helper Function: Give Hints For Project Names
}
