% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pubs_by_author.R
\name{plot_pubs_by_author}
\alias{plot_pubs_by_author}
\title{Plot Number of Publications by Author}
\usage{
plot_pubs_by_author(pubs_by_author_df)
}
\arguments{
\item{pubs_by_author_df}{a data frame with author names in column "value" and
number of publications (in column "n")}
}
\value{
plot of number of publications per author
}
\description{
Plot Number of Publications by Author
}
\examples{
refs_by_author_lastfirst <- create_df_from_endnote_xml() \%>\%
dplyr::filter(.data$key2 == "authors") \%>\%
dplyr::count(.data$value)  \%>\%
dplyr::arrange(dplyr::desc(.data$n))
plot_pubs_by_author(refs_by_author_lastfirst[1:30, ])
}
