% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pubs_by_year.R
\name{plot_pubs_by_year}
\alias{plot_pubs_by_year}
\title{Plot Publications By Year}
\usage{
plot_pubs_by_year(refs_df)
}
\arguments{
\item{refs_df}{reference dataframe as retrieved by
kwb.endnote::create_references_df()}
}
\value{
plot with publications by year
}
\description{
Plot Publications By Year
}
\examples{
endnote_list <- kwb.endnote::create_endnote_list()
refs_df <- kwb.endnote::create_references_df(endnote_list)
plot_pubs_by_year(refs_df)

}
