% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wordcloud_keywords.R
\name{plot_wordcloud_keywords}
\alias{plot_wordcloud_keywords}
\title{Plot Wordcloud Keywords}
\usage{
plot_wordcloud_keywords(keywords_df, ...)
}
\arguments{
\item{keywords_df}{keywords dataframe as retrieved by
create_keywords_df()}

\item{...}{additional arguments passed to wordcloud2::wordcloud2()}
}
\value{
wordcloud keywords plot
}
\description{
Plot Wordcloud Keywords
}
