% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_references_df_from_list.R
\name{record_list_to_df}
\alias{record_list_to_df}
\title{Reference List to Data Frame}
\usage{
record_list_to_df(record_list)
}
\arguments{
\item{record_list}{list with one record of create_endnote_list()}
}
\value{
data frame for record
}
\description{
Reference List to Data Frame
}
