% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{tidy_df}
\alias{tidy_df}
\title{Helper function: tidy dataframe}
\usage{
tidy_df(df, exclude_cols = "rec_number")
}
\arguments{
\item{df}{data frame as retrieved by create_references_df() or
clean_references_df()}

\item{exclude_cols}{vector of column names to exclude for gathering
(default: "rec_number")}
}
\value{
a tidy dataframe with columns rec_number, key and value
}
\description{
Helper function: tidy dataframe
}
\examples{
\dontrun{
endnote_list <- create_endnote_list()
refs_df <- create_references_df(endnote_list)
refs_df_tidy <- tidy_df(refs_df)
}
}
