% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{tidy_multi_cols}
\alias{tidy_multi_cols}
\title{Helper Function: Tidy Multi Cols}
\usage{
tidy_multi_cols(df)
}
\arguments{
\item{df}{as retrieved by create_references_df() or clean_references_df()}
}
\value{
tidy dataframe for all multi cols in df
}
\description{
Helper Function: Tidy Multi Cols
}
