% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{tidy_multi_cols_df}
\alias{tidy_multi_cols_df}
\title{Helper Function: Tidy Multi Cols Dataframe}
\usage{
tidy_multi_cols_df(df)
}
\arguments{
\item{df}{as retrieved by create_references_df() or clean_references_df()}
}
\value{
tidy dataframe for all multi cols in df
}
\description{
Helper Function: Tidy Multi Cols Dataframe
}
\examples{
\dontrun{endnote_list <- create_endnote_list()
refs_clean_df <- clean_references_df(endnote_list)
multi_cols_df <- tidy_multi_cols_df(refs_clean_df)
}
}
