% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{tidy_multi_cols_list}
\alias{tidy_multi_cols_list}
\title{Helper Function: Tidy Multi Cols List}
\usage{
tidy_multi_cols_list(df)
}
\arguments{
\item{df}{as retrieved by create_references_df() or clean_references_df()}
}
\value{
tidy list with a sublist for echa multi col in df (see:
get_available_multi_cols())
}
\description{
Helper Function: Tidy Multi Cols List
}
\examples{
\dontrun{endnote_list <- create_endnote_list()
refs_clean_df <- clean_references_df(endnote_list)
multi_cols_list <- tidy_multi_cols_list(refs_clean_df)
}
}
