% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{tidy_selected_cols}
\alias{tidy_selected_cols}
\title{Helper Function: Tidy Selected Cols}
\usage{
tidy_selected_cols(df, column = "author")
}
\arguments{
\item{df}{as retrieved by create_references_df() or clean_references_df()}

\item{column}{a multi col column e.g. "author"}
}
\value{
tidy dataframe for selected multi col in df
}
\description{
Helper Function: Tidy Selected Cols
}
