% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_clean_references_df_to_xlsx.R
\name{write_clean_references_df_to_xlsx}
\alias{write_clean_references_df_to_xlsx}
\title{Write Clean References Dataframe to XLSX}
\usage{
write_clean_references_df_to_xlsx(refs_df,
  file = "references_clean.xlsx", ...)
}
\arguments{
\item{refs_df}{references dataframe as retrieved by create_references_df()}

\item{file}{file where to save (default: "references_clean.xlsx")}

\item{...}{additional arguments passed to openxlsx::write.xlsx()
publication type}
}
\description{
Write Clean References Dataframe to XLSX
}
\examples{
\dontrun{endnote_list <- create_endnote_list()
refs_df <- create_references_df(endnote_list)
write_clean_references_df_to_xlsx(refs_df)
}
}
