% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_problematic_entries.R
\name{check_problematic_entries}
\alias{check_problematic_entries}
\title{Check Endnote for Problematic References}
\usage{
check_problematic_entries(endnote_list, replace_na = TRUE, dbg = TRUE)
}
\arguments{
\item{endnote_list}{list created with create_endnote_list()}

\item{replace_na}{if TRUE NAs are replaced with 'replace_na_text' (
default: TRUE) defined in relevant helper function}

\item{dbg}{show debug messages (default: TRUE)}
}
\value{
a data frame with problematic entries
}
\description{
Check Endnote for Problematic References
}
\examples{
\dontrun{endnote_list <- create_endnote_list()
problematic_entries <- check_problematic_entries(endnote_list)
head(problematic_entries)
}
}
