% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_references_df.R
\name{clean_accessibility}
\alias{clean_accessibility}
\title{Helper function: clean access information}
\usage{
clean_accessibility(access, replace_na = FALSE,
  replace_na_text = "add_public_or_confidential", dbg = TRUE)
}
\arguments{
\item{access}{vector with accessibility information}

\item{replace_na}{if TRUE NAs are replaced with 'replace_na_text' (
default: FALSE)}

\item{replace_na_text}{text for replacing NAs (default: "add_public_or_confidential")}

\item{dbg}{show debug messages (default: TRUE)}
}
\value{
vector with cleaned accessibility information
}
\description{
Helper function: clean access information
}
