% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_references_df.R
\name{clean_author_names}
\alias{clean_author_names}
\title{Helper function: clean author names}
\usage{
clean_author_names(author_names, replace_na = FALSE,
  replace_na_text = "add_author_lastname_semicolon_firstname",
  dbg = TRUE)
}
\arguments{
\item{author_names}{with author names to clean}

\item{replace_na}{if TRUE NAs are replaced with 'replace_na_text' (
default: FALSE)}

\item{replace_na_text}{text for replacing NAs (default:
"add_author_lastname_semicolon_firstname"")}

\item{dbg}{show debug messages (default: TRUE)}
}
\value{
vector with cleaned author names
}
\description{
Helper function: clean author names
}
