% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_references_df.R
\name{clean_dois}
\alias{clean_dois}
\title{Helper function: clean DOIs}
\usage{
clean_dois(dois)
}
\arguments{
\item{dois}{vectors with DOIs to clean}
}
\value{
cleaned DOIs
}
\description{
Helper function: clean DOIs
}
