% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_references_df.R
\name{clean_project_names}
\alias{clean_project_names}
\title{Helper function: clean project names}
\usage{
clean_project_names(project_names)
}
\arguments{
\item{vector}{with project names to clean}
}
\value{
vector with cleaned project names
}
\description{
Helper function: clean project names
}
