% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_references_df_from_list.R
\name{clean_references_df}
\alias{clean_references_df}
\title{Clean References Dataframe}
\usage{
clean_references_df(references_df)
}
\arguments{
\item{references_df}{as retrieved by create_references_df()}
}
\value{
cleaned references_df
}
\description{
Clean References Dataframe
}
