% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_endnote_list.R
\name{create_endnote_list}
\alias{create_endnote_list}
\title{Create List From Endnote XML}
\usage{
create_endnote_list(endnote_xml = default_xml())
}
\arguments{
\item{endnote_xml}{path to Endnote library exported as .xml (default:
\code{\link{default_xml()}})}
}
\value{
list of imported endnote_xml
}
\description{
Create List From Endnote XML
}
\examples{
endnote_list <- create_endnote_list()
str(endnote_list[1]$record)
}
