% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_references_df_from_list.R
\name{get_abstract}
\alias{get_abstract}
\title{Helper function: get abstract from list for a reference}
\usage{
get_abstract(record_list)
}
\arguments{
\item{record_list}{list with one record of create_endnote_list()}
}
\value{
one row abstract data frame
}
\description{
Helper function: get abstract from list for a reference
}
