% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KWB_documents.R
\name{get_reference_type_names}
\alias{get_reference_type_names}
\title{Helper function: get reference type names}
\usage{

  get_reference_type_names(endnote_xml = system.file("extdata/KWB_documents.xml",
  package = "kwb.endnote"))
}
\arguments{
\item{endnote_xml}{path to Endnote library exported as .xml (default:
system.file("extdata/KWB_documents.xml", package = "kwb.endnote")}
}
\value{
data.frame with columns record_id, rec_number, ref_type_id, ref_type_name
}
\description{
Helper function: get reference type names
}
\examples{
endnote_xml <- system.file("extdata/KWB_documents.xml",
package = "kwb.endnote")
ref_type_names <- get_reference_type_names(endnote_xml)
head(ref_type_names)
}
