% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_references_df_from_list.R
\name{get_secondary_authors}
\alias{get_secondary_authors}
\title{Helper function: get secondary authors from list for a reference}
\usage{
get_secondary_authors(record_list, collapse = FALSE)
}
\arguments{
\item{record_list}{list with one record of create_endnote_list()}

\item{collapse}{should separate fields in "style" be collapsed to one field?
(default: FALSE)}
}
\value{
one row authors data frame
}
\description{
Helper function: get secondary authors from list for a reference
}
