% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_references_df_from_list.R
\name{get_tertiary_authors}
\alias{get_tertiary_authors}
\title{Helper function: get tertiary authors from list for a reference}
\usage{
get_tertiary_authors(record_list)
}
\arguments{
\item{record_list}{list with one record of create_endnote_list()}
}
\value{
one row authors data frame
}
\description{
Helper function: get tertiary authors from list for a reference
}
