% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_clean_references_df_to_xlsx.R
\name{write_clean_references_df_to_xlsx}
\alias{write_clean_references_df_to_xlsx}
\title{Write Clean References Dataframe to XLSX}
\usage{
write_clean_references_df_to_xlsx(
  endnote_list,
  file = default_clean_xlsx(endnote_list),
  export_dir = ".",
  give_hints = FALSE,
  dbg = TRUE,
  ...
)
}
\arguments{
\item{endnote_list}{list created with create_endnote_list()}

\item{file}{name of file to save (default: default_clean_xlsx(endnote_list))}

\item{export_dir}{directory where to save 'file' (default: ".")}

\item{give_hints}{if TRUE hints will be generated, e.g.
"add_public_or_confidential" for accessiblity data}

\item{dbg}{show debug messages (default: TRUE)}

\item{...}{additional arguments passed to openxlsx::write.xlsx()
publication type}
}
\description{
Write Clean References Dataframe to XLSX
}
\examples{
\dontrun{endnote_list <- create_endnote_list()
write_clean_references_df_to_xlsx(endnote_list)
}
}
