% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_references_df_to_xlsx.R
\name{write_references_df_to_xlsx}
\alias{write_references_df_to_xlsx}
\title{Write References Dataframe to XLSX}
\usage{
write_references_df_to_xlsx(refs_df, file = "references.xlsx", ...)
}
\arguments{
\item{refs_df}{references dataframe as retrieved by create_references_df()}

\item{file}{file where to save (default: "references.xlsx")}

\item{...}{additional arguments passed to openxlsx::write.xlsx()}
}
\value{
write references dataframe to xlsx with one sheet for each
publication type
}
\description{
Write References Dataframe to XLSX
}
\examples{
\dontrun{endnote_list <- create_endnote_list()
refs_df <- create_references_df(endnote_list)
write_references_df_to_xlsx(refs_df)
}
}
