% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_references_df_to_xlsx.R
\name{write_references_df_to_xlsx}
\alias{write_references_df_to_xlsx}
\title{Write References Dataframe to XLSX}
\usage{
write_references_df_to_xlsx(
  endnote_list,
  file = default_xlsx(endnote_list),
  export_dir = ".",
  dbg = TRUE,
  ...
)
}
\arguments{
\item{endnote_list}{list created with create_endnote_list()}

\item{file}{name of file to save (default: default_xlsx(endnote_list))}

\item{export_dir}{directory where to save 'file' (default: ".")}

\item{dbg}{show debug messages (default: TRUE)}

\item{...}{additional arguments passed to openxlsx::write.xlsx()}
}
\value{
write references dataframe to xlsx with one sheet for each
publication type
}
\description{
Write References Dataframe to XLSX
}
\examples{
\dontrun{endnote_list <- create_endnote_list()
write_references_df_to_xlsx(endnote_list)
}
}
